+++ flowcept/commons/doc_db/document_db_dao.py
@@ -1,4 +1,3 @@
-import logging
 from typing import List, Dict
 from bson import ObjectId
 from pymongo import MongoClient, UpdateOne
@@ -10,7 +9,7 @@ from flowcept.configs import (
     MONGO_DB,
     MONGO_COLLECTION,
 )
-from flowcept.flowcept_consumer.consumer_utils import (
+from flowcept.commons.consumer_utils import (
     curate_dict_task_messages,
 )
