+++ flowcept/flowceptor/plugins/tensorboard/tensorboard_interceptor.py
@@ -97,13 +97,3 @@ class TensorboardInterceptor(BaseInterceptor):
         )
         observer.start()
         self.logger.debug(f"Watching {self.settings.file_path}")
-
-
-if __name__ == "__main__":
-    try:
-        interceptor = TensorboardInterceptor()
-        interceptor.observe()
-        while True:
-            time.sleep(interceptor.settings.watch_interval_sec)
-    except KeyboardInterrupt:
-        sys.exit(0)
